requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
kombu==4.6.7
Markdown==3.2.1
numpy==1.18.1
pandas==1.0.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.8
vine==1.3.0
xlrd==1.2.0
